Create new partial class for map loading
